+++ doc/source/whatsnew.rst
@@ -2,7 +2,7 @@
 What's New
 ======================
 
-v2.1.0.dev0
+v2.1.0
 ===========
 
 * Added the ``cffi-gen-src`` command-line tool (also invocable as